fix: Corrected module package names
