+++ activiti-cloud-services-runtime-bundle/messages-events/src/main/java/org/activiti/cloud/services/messages/events/producer/StartMessageDeployedEventMessageProducer.java
@@ -1,26 +1,13 @@
-/*
- * Copyright 2019 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+package org.activiti.cloud.services.messages.events.producer;
 
-package org.activiti.cloud.services.message.events;
-
-import org.activiti.api.process.model.MessageSubscription;
+import org.activiti.api.process.model.StartMessageSubscription;
 import org.activiti.api.process.model.builders.MessageEventPayloadBuilder;
-import org.activiti.api.process.model.events.MessageSubscriptionCancelledEvent;
+import org.activiti.api.process.model.events.StartMessageDeployedEvent;
 import org.activiti.api.process.model.payloads.MessageEventPayload;
 import org.activiti.api.process.runtime.events.listener.ProcessRuntimeEventListener;
+import org.activiti.cloud.services.messages.events.MessageEventHeaders;
+import org.activiti.cloud.services.messages.events.support.MessageSenderTransactionSynchronization;
+import org.activiti.cloud.services.messages.events.support.StartMessageDeployedEventMessageBuilderFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.lang.NonNull;
@@ -28,34 +15,36 @@ import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
-public class MessageSubscriptionCancelledEventMessageProducer implements ProcessRuntimeEventListener<MessageSubscriptionCancelledEvent> {
+public class StartMessageDeployedEventMessageProducer implements ProcessRuntimeEventListener<StartMessageDeployedEvent> {
 
-    private static final Logger logger = LoggerFactory.getLogger(MessageSubscriptionCancelledEventMessageProducer.class);
+    private static final Logger logger = LoggerFactory.getLogger(BpmnMessageSentEventMessageProducer.class);
 
-    private final MessageSubscriptionEventMessageBuilderFactory messageBuilderFactory;
+    private final StartMessageDeployedEventMessageBuilderFactory messageBuilderFactory;
     private final MessageChannel messageChannel;
 
-    public MessageSubscriptionCancelledEventMessageProducer(@NonNull MessageChannel messageChannel,
-                                                            @NonNull MessageSubscriptionEventMessageBuilderFactory messageBuilderFactory) {
+    public StartMessageDeployedEventMessageProducer(@NonNull MessageChannel messageChannel,
+                                                    @NonNull StartMessageDeployedEventMessageBuilderFactory messageBuilderFactory) {
         this.messageChannel = messageChannel;
         this.messageBuilderFactory = messageBuilderFactory;
     }
-
+    
     @Override
-    public void onEvent(@NonNull MessageSubscriptionCancelledEvent event) {
+    public void onEvent(StartMessageDeployedEvent event) {
         logger.debug("onEvent: {}", event);
-        
+
         if (!TransactionSynchronizationManager.isSynchronizationActive()) {
             throw new IllegalStateException("requires active transaction synchronization");
         }
-
-        MessageSubscription messageSubscription = event.getEntity();
+        
+        StartMessageSubscription messageSubscription = event.getEntity()
+                                                            .getMessageSubscription();
 
         MessageEventPayload messageEventPayload = MessageEventPayloadBuilder.messageEvent(messageSubscription.getEventName())
                                                                             .withCorrelationKey(messageSubscription.getConfiguration())
                                                                             .build();
         
-        Message<MessageEventPayload> message = messageBuilderFactory.create(event.getEntity())
+        
+        Message<MessageEventPayload> message = messageBuilderFactory.create(event)
                                                                     .withPayload(messageEventPayload)
                                                                     .setHeader(MessageEventHeaders.MESSAGE_EVENT_TYPE,
                                                                                event.getEventType()
@@ -64,7 +53,7 @@ public class MessageSubscriptionCancelledEventMessageProducer implements Process
 
         TransactionSynchronizationManager.registerSynchronization(new MessageSenderTransactionSynchronization(message,
                                                                                                               messageChannel));
+        
     }
-
 
 }
