+++ activiti-cloud-services-runtime-bundle/messages-events/src/main/resources/META-INF/spring.factories
