+++ activiti-cloud-services-runtime-bundle/messages-events/src/main/resources/config/message-events-channels.properties
