+++ activiti-cloud-services-runtime-bundle/pom.xml
@@ -22,7 +22,7 @@
     <module>activiti-cloud-services-identity-basic</module>
     <module>activiti-cloud-services-identity-keycloak</module>
     <module>activiti-cloud-services-job-executor</module>
-    <module>activiti-cloud-services-messages-events</module>
+    <module>messages-events</module>
     <module>messages-aggregator</module>
   </modules>
 </project>
