+++ activiti-cloud-starter-runtime-bundle/src/test/java/org/activiti/cloud/starter/tests/runtime/MessageEventsIT.java
@@ -43,11 +43,11 @@ import org.activiti.cloud.api.model.shared.CloudVariableInstance;
 import org.activiti.cloud.api.process.model.CloudProcessInstance;
 import org.activiti.cloud.services.core.commands.ReceiveMessageCmdExecutor;
 import org.activiti.cloud.services.core.commands.StartMessageCmdExecutor;
-import org.activiti.cloud.services.message.events.BpmnMessageReceivedEventMessageProducer;
-import org.activiti.cloud.services.message.events.BpmnMessageSentEventMessageProducer;
-import org.activiti.cloud.services.message.events.BpmnMessageWaitingEventMessageProducer;
-import org.activiti.cloud.services.message.events.MessageSubscriptionCancelledEventMessageProducer;
-import org.activiti.cloud.services.message.events.StartMessageDeployedEventMessageProducer;
+import org.activiti.cloud.services.messages.events.producer.BpmnMessageReceivedEventMessageProducer;
+import org.activiti.cloud.services.messages.events.producer.BpmnMessageSentEventMessageProducer;
+import org.activiti.cloud.services.messages.events.producer.BpmnMessageWaitingEventMessageProducer;
+import org.activiti.cloud.services.messages.events.producer.MessageSubscriptionCancelledEventMessageProducer;
+import org.activiti.cloud.services.messages.events.producer.StartMessageDeployedEventMessageProducer;
 import org.activiti.cloud.starter.tests.helper.ProcessInstanceRestTemplate;
 import org.activiti.engine.RuntimeService;
 import org.awaitility.Awaitility;
